Recording refactor
