+++ src/kz/recording/kz_recording_buffer.cpp
@@ -1,0 +1,132 @@
+#include "kz_recording.h"
+#include "kz/language/kz_language.h"
+extern CConVar<bool> kz_replay_recording_debug;
+
+// CircularRecorder method implementations
+void CircularRecorder::TrimOldCommands(u32 currentTick)
+{
+	i32 numToRemove = 0;
+	for (i32 i = 0; i < this->cmdData->GetReadAvailable(); i++)
+	{
+		CmdData data;
+		if (!this->cmdData->Peek(&data, 1, i))
+		{
+			break;
+		}
+		if (data.serverTick + 2 * 60 * 64 < currentTick)
+		{
+			numToRemove++;
+		}
+		else
+		{
+			break;
+		}
+	}
+	this->cmdData->Advance(numToRemove);
+	this->cmdSubtickData->Advance(numToRemove);
+}
+
+void CircularRecorder::TrimOldWeaponEvents(u32 currentTick)
+{
+	i32 numToRemove = 0;
+	for (i32 i = 0; i < this->weaponChangeEvents->GetReadAvailable(); i++)
+	{
+		WeaponSwitchEvent data;
+		if (!this->weaponChangeEvents->Peek(&data, 1, i))
+		{
+			break;
+		}
+		if (data.serverTick + 2 * 60 * 64 < currentTick)
+		{
+			this->earliestWeapon = data.econInfo;
+			numToRemove++;
+			continue;
+		}
+		break;
+	}
+	this->weaponChangeEvents->Advance(numToRemove);
+}
+
+void CircularRecorder::TrimOldRpEvents(u32 currentTick)
+{
+	i32 numToRemove = 0;
+	for (i32 i = 0; i < this->rpEvents->GetReadAvailable(); i++)
+	{
+		RpEvent event;
+		if (!this->rpEvents->Peek(&event, 1, i))
+		{
+			break;
+		}
+		if (event.serverTick + 2 * 60 * 64 < currentTick)
+		{
+			switch (event.type)
+			{
+				case RPEVENT_MODE_CHANGE:
+				{
+					V_strncpy(this->earliestMode.value().name, event.data.modeChange.name, sizeof(this->earliestMode.value().name));
+					V_strncpy(this->earliestMode.value().md5, event.data.modeChange.md5, sizeof(this->earliestMode.value().md5));
+					break;
+				}
+				case RPEVENT_STYLE_CHANGE:
+				{
+					if (event.data.styleChange.clearStyles)
+					{
+						this->earliestStyles.value().clear();
+					}
+					RpModeStyleInfo style = {};
+					V_strncpy(style.name, event.data.styleChange.name, sizeof(style.name));
+					V_strncpy(style.md5, event.data.styleChange.md5, sizeof(style.md5));
+					this->earliestStyles.value().push_back(style);
+					break;
+				}
+				case RPEVENT_CHECKPOINT:
+				{
+					this->earliestCheckpointIndex = event.data.checkpoint.index;
+					this->earliestCheckpointCount = event.data.checkpoint.checkpointCount;
+					this->earliestTeleportCount = event.data.checkpoint.teleportCount;
+					break;
+				}
+			}
+			numToRemove++;
+			continue;
+		}
+		break;
+	}
+	this->rpEvents->Advance(numToRemove);
+}
+
+void CircularRecorder::TrimOldJumps(u32 currentTick)
+{
+	i32 numToRemove = 0;
+	for (i32 i = 0; i < this->jumps->GetReadAvailable(); i++)
+	{
+		RpJumpStats *jump = this->jumps->PeekSingle(i);
+		if (!jump)
+		{
+			break;
+		}
+		if (jump->overall.serverTick + 2 * 60 * 64 < currentTick)
+		{
+			numToRemove++;
+			continue;
+		}
+		break;
+	}
+	this->jumps->Advance(numToRemove);
+}
+
+void KZRecordingService::WriteCircularBufferToFile(f32 duration, const char *cheaterReason)
+{
+	if (strlen(cheaterReason) > 0)
+	{
+		CheaterRecorder recorder(this->player, cheaterReason);
+		recorder.WriteToFile();
+		META_CONPRINTF("[KZ] Cheater replay saved: %s\n", recorder.uuid.ToString().c_str());
+	}
+	else
+	{
+		ManualRecorder recorder(this->player, duration);
+		recorder.WriteToFile();
+		player->languageService->PrintChat(true, false, "Replay - Manual Replay Saved", recorder.uuid.ToString().c_str());
+	}
+}
